refactor: PR fixes
