--- src/main/java/io/gravitee/policy/generatehttpsignature/configuration/SchemeTypeConfiguration.java
@@ -1,39 +1,0 @@
-/*
- * Copyright © 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.policy.generatehttpsignature.configuration;
-
-import java.util.ArrayList;
-import java.util.List;
-import lombok.*;
-
-/**
- * @author Brent HUNTER (brent.hunter at graviteesource.com)
- * @author GraviteeSource Team
- */
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SchemeTypeConfiguration {
-
-    private boolean enabled;
-
-    // Optional - List of additional headers to add to the Webhook signature creation
-    private String headersDelimiter;
-
-    private List<String> headers = new ArrayList<>();
-}
