+++ src/main/resources/plugin.properties
@@ -6,5 +6,5 @@ class=io.gravitee.policy.generatehttpsignature.GenerateHttpSignaturePolicy
 type=policy
 category=security
 icon=generate-http-signature.svg
-proxy=REQUEST
-message=REQUEST,RESPONSE,MESSAGE_RESPONSE
+http_proxy=REQUEST,RESPONSE
+http_message=REQUEST,RESPONSE,SUBSCRIBE
