+++ src/test/java/io/gravitee/policy/generatehttpsignature/v3/GenerateHttpSignaturePolicyV4EmulationIntegrationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.policy.generatehttpsignature.v2;
+package io.gravitee.policy.generatehttpsignature.v3;
 
 import io.gravitee.apim.gateway.tests.sdk.annotations.GatewayTest;
 import io.gravitee.definition.model.ExecutionMode;
